Added support of "NOCONFIGURE" environmental variable. Fixes to code according to shellcheck.
This is a port of changes in NetworkManager-openvpn (https://github.com/NetworkManager/NetworkManager-openvpn/pull/3)
to NetworkManager-pptp.

Example usage:
in RPM spec we do not want to call `./confiogure`, we want to use
a special macro, e.g. %configure or %configure2_5x,
which will automatically set different flags and options.
So, we take a pure git tarball and run:
`env NOCONFIGURE=autogen.sh`
`%configure2_5x`
https://abf.io/mikhailnov/networkmanager-openvpn/blob/rosa2016.1/networkmanager-openvpn.spec#lc-63